Refactored ListPages
